mv pr_plan tests to sanity and fix imports
